+++ smok/extras/pred_database/__init__.py
@@ -1,0 +1,3 @@
+from .base import BasePredicateDatabase
+from .pickling import PicklingPredicateDatabase
+