+++ smok/threads/communicator.py
@@ -93,6 +93,7 @@ class CommunicatorThread(TerminableThread):
             raise
 
         predicates_found = set()
+        preds_to_dump = []
         for predicate_dict in resp:
             predicate_id = predicate_dict['predicate_id']
             predicates_found.add(predicate_id)
@@ -126,10 +127,14 @@ class CommunicatorThread(TerminableThread):
                     if stat.group != group:
                         stat.on_group_changed(group)
 
-            predicates_to_delete = set(self.device.predicates.keys()) - predicates_found
-            for predicate_id in predicates_to_delete:
-                self.device.predicates[predicate_id].on_offline()
-                del self.device.predicates[predicate_id]
+                predicate_dict['silencing'] = silencing
+                preds_to_dump.append(predicate_dict)
+
+        predicates_to_delete = set(self.device.predicates.keys()) - predicates_found
+        for predicate_id in predicates_to_delete:
+            self.device.predicates[predicate_id].on_offline()
+            del self.device.predicates[predicate_id]
+        self.device.pred_database.set_new_predicates(preds_to_dump)
 
         if not self.last_predicates_synced:
             self.device.ready_lock.release()
@@ -176,7 +181,7 @@ class CommunicatorThread(TerminableThread):
             if e.is_no_link():
                 self.device.on_failed_sync()
             if not e.is_clients_fault():
-                logger.debug(f'Failed to sync sensor writes: {e}', e, exc_info=e)
+                logger.debug('Failed to sync sensor writes: %s', e, exc_info=e)
                 sync.nack()
                 raise
             else: